Retrieve customer feature is working

+++ src/main/java/com/example/demo/CustomerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customers, Integer>  {
+public interface CustomerRepository extends JpaRepository<Customers,Long>  {
 
 }
